Added new flow functionality to add-another apprenticeship route

+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.ProviderCommitments.Web/Models/ReservationsAddDraftApprenticeshipRequest.cs
@@ -6,7 +6,7 @@ namespace SFA.DAS.ProviderCommitments.Web.Models
 {
     public class ReservationsAddDraftApprenticeshipRequest : BaseReservationsAddDraftApprenticeshipRequest, IAuthorizationContextModel
     {
-        public int ProviderId { get; set; }
+        public long ProviderId { get; set; }
         public string CohortReference { get; set; }
         public long? CohortId { get; set; }
     }
